+++ src/test/fuzz/block_index.cpp
@@ -12,6 +12,8 @@
 #include <txdb.h>
 #include <validation.h>
 
+using kernel::CBlockFileInfo;
+
 namespace {
 
 const BasicTestingSetup* g_setup;
